+++ src/main/java/io/github/scaredsmods/potion_totems/event/ClientModEvents.java
@@ -3,6 +3,7 @@ package io.github.scaredsmods.potion_totems.event;
 
 import io.github.scaredsmods.potion_totems.init.PTItems;
 import io.github.scaredsmods.potion_totems.init.PTMenuTypes;
+import io.github.scaredsmods.potion_totems.screen.AdvancedInfuserScreen;
 import io.github.scaredsmods.potion_totems.screen.InfuserScreen;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.util.FastColor;
@@ -12,7 +13,6 @@ import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.RegisterColorHandlersEvent;
 import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
-import net.neoforged.neoforge.registries.NewRegistryEvent;
 
 @EventBusSubscriber(bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public class ClientModEvents {
@@ -33,8 +33,8 @@ public class ClientModEvents {
 
     @SubscribeEvent
     public static void registerScreens(RegisterMenuScreensEvent event) {
-
         event.register(PTMenuTypes.INFUSER_MENU.get(), InfuserScreen::new);
+        event.register(PTMenuTypes.ADVANCED_INFUSER_MENU.get(), AdvancedInfuserScreen::new);
     }