+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+
+# 1.1.0
+## Features
+✅ Added an advanced infuser block
+
+✅ Added a villager with predefined potion (totems)
+
+✅ Added a custom potion Aggression
+
+## Changes
+🔄 Potion Totems now requires Fzzy Config instead of regular NeoForge config system
+
+## Fixes
+🛠️ Readded vanilla effects to the totem.
+
+🛠️ Removed occlusion on the regular infuser
+
+️🛠️ Fixed the shape of the infuser model by adding voxel shapes
+
+## Removals
+❌ Empty
+
+
 # 1.0.2
 - Added a crafting recipe for the infuser